requirements.txt
diffusers
torch
accelerate
gradio
Pillow
numpy
matplotlib
huggingface_hub
transformers
python-dotenv
flask
flask-cors